removed attempt to utilize leftover points from parallel computation
This commit reverses some previous changes where I tried to implement a way to utilize the 'leftover' points at the end of the parallel processing step. @aarchiba discovered the behavior actually is quite buggy, and the fix involves enough additional elements injected into the code that for simplicity I reverted this change rather than trying to catch the relevant exceptions.