mcmc bugfix and pool exception
Fixed a bug where for the `'classic'` sampler it would automatically fail because the `scale` variable for the proposals wasn't properly transferred to the  function.

Changed behavior where if users do not supply their own pool a `ValueException` is raised. Also changed the variable from `n_threads` to `queue_size`.